test: 만료된 토큰 확인 테스트 클래스 분ㄴ리 (#222)
