+++ src/test/java/com/exchangediary/member/service/JwtServiceTest.java
@@ -1,14 +1,12 @@
 package com.exchangediary.member.service;
 
-import com.exchangediary.global.exception.serviceexception.UnauthorizedException;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, properties = {"security.jwt.expiration-time=1000"})
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class JwtServiceTest {
     @Autowired
     private JwtService jwtService;
@@ -29,16 +27,5 @@ public class JwtServiceTest {
         String token = jwtService.generateToken(memberId);
 
         jwtService.verifyToken(token);
-    }
-
-    @Test
-    void 만료된_토큰_검증() throws InterruptedException {
-        Long memberId = 1L;
-        String token = jwtService.generateToken(memberId);
-
-        Thread.sleep(1000);
-        assertThrows(UnauthorizedException.class, () ->
-            jwtService.verifyToken(token)
-        );
     }
 }
